Our clock constraints for Quartus were totally broken.  This commit fixes them.
1) Moved Quartus compilation to the compileDirectory.  quartus_sta
assumes all files in the qsf are relative to the directory in which
the qsf file resides. Boo.  We haven't been loading clock constrains
correctly for some time.

2) Added a check for timing failure.  We now print out a message
stating whether we meet or miss timing.

3) Added clock crossing annotations for JTAG device.

The end result of these commits is that we can run multiple clock
domains codes compiled through Quartus. Yay!